backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.1.0
anthropic==0.50.0
python-dotenv==1.0.0
PyMuPDF==1.23.8
sqlalchemy==2.0.27
psycopg2-binary==2.9.9 